Import shared folders (rsync server)
As the core rsync server only transfers numeric ids, we need an
additional rsync server that can resolve user and group names, so
owners, permissions and ACLs are preserved by the rsync
transaction. This server is used as the last rsync pass to synchronize
ACLs. The core rsync server is used to transfer the data
payload in previous passes.