api. Return the "file_server" hint for UI
The UI needs to known if the DC that it is configuring can be used as a
LAN SMB file server (and AD logon server too...). The `get-defaults`
action now returns an attribute "can_install_file_server" that checks if
the node and the AD domain allows it according to our business rules.